feat: 조회 결과 DTO 페이지 정보 포함
