+++ src/main/java/com/fastcampus/reserve/domain/order/dto/response/OrderHistoriesDto.java
@@ -1,29 +1,35 @@
 package com.fastcampus.reserve.domain.order.dto.response;
 
 import com.fastcampus.reserve.domain.order.Order;
-import com.fastcampus.reserve.domain.order.orderitem.OrderItem;
-import java.time.LocalDate;
 import java.util.List;
 import lombok.Builder;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 
 @Builder
 public record OrderHistoriesDto(
-        Long orderId,
-        LocalDate reserveDate,
-        List<OrderItemInfoDto> orderItems
+        int size,
+        int pageNumber,
+        int totalPages,
+        long totalElements,
+        List<OrderHistoryDto> orderHistories
 ) {
 
-    public static OrderHistoriesDto from(Order order) {
+    public static OrderHistoriesDto from(Page<Order> orders) {
+        Pageable pageable = orders.getPageable();
+
         return OrderHistoriesDto.builder()
-                .orderId(order.getId())
-                .reserveDate(order.getCreatedDate().toLocalDate())
-                .orderItems(getOrderItems(order.getOrderItems()))
+                .size(pageable.getPageSize())
+                .pageNumber(pageable.getPageNumber())
+                .totalPages(orders.getTotalPages())
+                .totalElements(orders.getTotalElements())
+                .orderHistories(getOrderHistories(orders.getContent()))
                 .build();
     }
 
-    private static List<OrderItemInfoDto> getOrderItems(List<OrderItem> order) {
-        return order.stream()
-                .map(OrderItemInfoDto::from)
+    private static List<OrderHistoryDto> getOrderHistories(List<Order> orders) {
+        return orders.stream()
+                .map(OrderHistoryDto::from)
                 .toList();
     }
 }
